pyproject.toml
[tool.poetry]
name = "automatedgreythr"
version = "0.1.0"
description = "This aplication automates the process of GreytHR sign-in and sign-out"
authors = ["Pratik Ghodke <pratikghodke41@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.3"
pydantic-settings = "^2.0.3"
flask-restful = "^0.3.10"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.0.5"
flask-pydantic = "^0.11.0"
gunicorn = "^21.2.0"
cryptography = "^41.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
